Fixed event horizon regen, made active gates glow
Activating outgoing event horizons previously replaced most of the air
with event horizon blocks, to handle the case where you built inside the
gate and then removed what you built.  This now happens with both
incoming and outgoing connections, with everything in
#minecraft:replaceable, and in the exact area of the event horizon
instead of an approximation.

This is also now a part of animations:general_wormhole_animations,
because while I was at it I made the event horizons glow when active and
stop glowing when the gate turns off.