Use lectern instead of bat spawn egg for DHDs
This allows more intuitive block placement, especially in the case of
trying to place it on top of a redstone torch (useful for redial
functionality).  It also gets rid of every-tick spawn detection, which
probably improves performance but I havenʼt tested because that isnʼt
the main reason.

This fixes the appearence and sound of a suddenly spawning-and-dying
bat, but replaces it with the wrong block appearing for a tick.  I find
that less intrusive but it isnʼt a straight win.