Base travel cooldown on connection not time
Now you are in cooldown whenever interacting with the most recent
wormhole you went through, but not others.  This means if you wait just
one tick before going the wrong way through a *different* wormhole, you
die, but if you go through a wormhole you can spend all the rest of its
open time dancing in and out of the outbound horizon.  This is useful
for times when the cooldown timer detects you are not in a wormhole but
the damage command detects you are, or when you have a bad connection
and the server keeps changing where it thinks you are in response.

Additionally, wormhole connection IDs now work like DHD IDs; the limit
is 4 billion, not 8.