Make autodialing harder to accidentally DOS
Gates now know if they recently autodialed and required the power source
to be off for a tick before dialing again.  This means a torch under the
DHD is no longer a DOS on the destination gate.

Additionally, more power sources are now supported.  Ideally it would
detect any redstone activation but this is in fact hard.